refactor: main 함수 분리
