+++ src/basic/features/cart/index.js
@@ -1,4 +1,5 @@
 // Cart Feature 진입점
-export { calculateCartTotals, updateCartState, calculateItemData } from './business/calculations.js';
-export { updateCartItemStyles, updateCartSummary } from './ui/dom-updates.js';
-export { getCartProductTypes } from './helpers/cart-utils.js'; 
+export { calculateCartTotals, updateCartState, calculateItemData, calculateCompleteCartTotals } from './business/calculations.js';
+export { updateCartItemStyles, updateCartSummary, updatePricesInCart, updateCartTotalsDisplay } from './ui/dom-updates.js';
+export { getCartProductTypes } from './helpers/cart-utils.js';
+export { handleAddToCart, handleCartActions } from './handlers/cart-handlers.js';
