+++ src/basic/features/products/index.js
@@ -1,3 +1,3 @@
 // Products Feature 진입점
 export { initializeProducts, findProduct, getLowStockProducts, findProductById } from './data/product-manager.js';
-export { getOptionData, getProductDisplayInfo } from './ui/product-options.js'; 
+export { getOptionData, getProductDisplayInfo, updateSelectOptions } from './ui/product-options.js';
